requirements.txt
requests
beautifulsoup4
praw
python-telegram-bot==13.15
